preprocessing/requirements.txt
opencv-python>=4.5.0
numpy>=1.19.0
Pillow>=9.0.0
pathlib>=1.0.0
tqdm>=4.60.0
argparse>=1.4.0 